requirements.txt
backports.functools-lru-cache==1.5
beautifulsoup4==4.8.1
bs4==0.0.1
Pillow==6.2.1
pytesseract==0.3.0
soupsieve==1.9.4
requests==2.22.0
